requirements.txt
aiogram==3.13.1
#aiohttp==3.10.10
bs4==0.0.2
cache3==0.3.0
httpx==0.27.2
loguru==0.7.2
lxml==5.3.0
pydantic==2.9.2
tortoise-orm==0.21.6
vkbottle==4.4.2